Change api of the solver from fixed matrix<double> to parametrized abstract class, inherit existed matrix class from the abstract class. Eliminate unecessary copy of the data holder matrix into inner matrix. Migrate to header-only library (by the fact library is already header-only).
