+++ benchmarks/tests/main_google.cpp
@@ -17,7 +17,7 @@ static void BM_solve (benchmark::State & state)
 {
     state.PauseTiming ();
     if ( state.range_x () < matrices.size () ) {
-        Munkres munkres;
+        Munkres<double> munkres;
         while ( state.KeepRunning () ) {
             auto matrix = *matrices [state.range_x ()];
             state.ResumeTiming ();
