+++ benchmarks/tests/main_rdtsc.cpp
@@ -29,7 +29,7 @@ int main (int argc, char * argv [])
         for (size_t j = 0; j < runs; ++j) {
             uint64_t rdtscMinRun = std::numeric_limits<uint64_t>::max ();
             for (size_t k = 0; k < iterations; ++k) {
-                Munkres munkres;
+                Munkres<double> munkres;
                 auto matrix = *matrices [i];
                 uint64_t rdtsc = __rdtsc ();
                 munkres.solve (matrix);
